move POR classes to Library-TraceCheckerUtils
